short funciona, cavalo anda frente saida terminal

+++ movimentos.h
@@ -1,15 +1,15 @@
-int cimaDireita(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int cimaDireita(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
 
-int direitaCima(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int direitaCima(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
 
-int direitaBaixo(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int direitaBaixo(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
 
-int baixoDireita(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int baixoDireita(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
 
-int baixoEsquerda(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int baixoEsquerda(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
 
-int esquerdaBaixo(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int esquerdaBaixo(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
 
-int esquerdaCima(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int esquerdaCima(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
 
-int cimaEsquerda(int tabuleiro[][8], int* m, int* n, int* nMovs, short int* hist);
+int cimaEsquerda(short int tabuleiro[][8], short int* m, short int* n, int* nMovs, short int* hist);
